requirements.txt
opencv-python
mysql-connector-python
boto3
mediapipe
